ui: extract count value formatter
Release note: None
